fix(tray): make icon activation reliable on Windows 7

+++ src/window_procedure/window_message_handlers.c
@@ -99,7 +99,7 @@ LRESULT HandleDestroy(HWND hwnd, WPARAM wp, LPARAM lp) {
 }
 
 LRESULT HandleTrayIcon(HWND hwnd, WPARAM wp, LPARAM lp) {
-    HandleTrayIconMessage(hwnd, (UINT)wp, (UINT)lp);
+    HandleTrayIconMessage(hwnd, wp, lp);
     return 0;
 }
